# Lots of performance enhancements