--- a/src/Portable.Xaml/Portable.Xaml/XamlNodeListReader.cs
+++ b/src/Portable.Xaml/Portable.Xaml/XamlNodeListReader.cs
@@ -63,7 +63,8 @@ namespace Portable.Xaml
 
 		public override XamlNodeType NodeType
 		{
-			get { return node.NodeType; }
+			get { return node?.NodeType ?? XamlNodeType.None; }
+			//get { return node.NodeType; }
 		}
 
 		public override XamlSchemaContext SchemaContext