… src/Portable.Xaml/Portable.Xaml/XamlMember.cs
… src/Portable.Xaml/Portable.Xaml/XamlNode.cs
… src/Portable.Xaml/Portable.Xaml/XamlNodeQueue.cs
… src/Portable.Xaml/Portable.Xaml/XamlNodeQueueReader.cs
… src/Portable.Xaml/Portable.Xaml/XamlObjectNodeIterator.cs
… src/Portable.Xaml/Portable.Xaml/XamlObjectReader.cs
… src/Portable.Xaml/Portable.Xaml/XamlSchemaContext.cs
… src/Portable.Xaml/Portable.Xaml/XamlType.cs
… src/Portable.Xaml/Portable.Xaml/XamlWriterInternalBase.cs
… src/Portable.Xaml/Portable.Xaml/XamlXmlReader.cs
… src/Portable.Xaml/Portable.Xaml/XamlXmlWriter.cs