Now with these results:

## Load
Method |          Mean |     StdDev | Scaled | Scaled-StdDev |   Gen 0 | Allocated |
------------- |-------------- |----------- |------- |-------------- |-------- |---------- |
 PortableXaml |   897.6095 us | 24.2359 us |   1.00 |          0.00 | 19.7368 | 200.19 kB |
   SystemXaml | 1,427.0262 us | 52.7652 us |   1.59 |          0.07 |  3.0382 | 155.19 kB |

## Save
Method |        Mean |     StdDev | Scaled | Scaled-StdDev |   Gen 0 | Allocated |
------------- |------------ |----------- |------- |-------------- |-------- |---------- |
 PortableXaml | 990.8799 us | 18.2260 us |   1.00 |          0.00 | 25.0000 |    408 kB |
   SystemXaml | 923.7489 us | 21.5989 us |   0.93 |          0.03 |       - | 120.03 kB |